Moved events to dedicated package
